// src/app/features/performance-recording/typing/models/typing-trait.model.ts
export interface TypingTrait {
  typingTrait: number;
  createdBy: number;
  creationDate: string;
  effectiveFromDate: string;
  effectiveToDate: string;
  isActive: string;
  maxLimit: number;
  measurementUnitCd: number;
  minLimit: number;
  modifiedBy: number;
  modifiedDate: string;
  typingTraitDesc: string;
  measurementUnitDesc: string;
}